// frontend/src/types/enums/app-routes.enum.ts
export enum AppRoutes {
  Root = '/',
  Any = '*',
  SignIn = '/sign-in',
  SignUp = '/sign-up',
  Home = '/home',
  Portfolio = '/portfolio',
  DetailsProject = '/projects/:id',
  DetailsUser = '/users/:id',
  MessageCenter = '/chats',
  DetailsChat = '/chats/:id',
  About = '/about',
  Profile = '/profile',
  Dashboard = '/dashboard',
  TokenMarket = '/token-market',
}